pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "site-builder-eval-pack"
version = "0.1.0"
description = "Evaluation pack for site builder agent skills"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "boto3>=1.34",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "Pillow>=10.0",
    "scikit-image>=0.22",
    "playwright>=1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "moto[bedrock]>=5.0",
]

[project.scripts]
eval-cli = "eval_packs.local.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["eval_core", "eval_packs"]
